requirements.txt
gTTS==2.3.2
Pillow==9.4.0
playsound==1.3.0
streamlit==1.32.2